requirements.txt
requests
marshmallow
python-dotenv

# Database
psycopg2-binary
sqlalchemy